+++ include/hw/misc/memory_backend/mem_fault.h
@@ -19,8 +19,8 @@
 //  `pulldowns` is a bitfield corresponding to the pins that should be shortcut
 //     to LOW -> active on 1. 
 typedef struct {
-    uint8_t dq_pullups;
-    uint8_t dq_pulldowns;
+    uint64_t dq_pullups;
+    uint64_t dq_pulldowns;
 
     uint16_t a_pullups;
     uint16_t a_pulldowns;
